Fixes bug regarding themes with spaces in filename
The theme detection loop was not properly handling folders that had a space in their name. This caused a bug where it would ignore any and all characters after a space, which would cause the copy operation and menu rendering for that theme to fail. This is fixed by setting the IFS variable to tell bash to only split on newlines and to not also split on spaces.